Remove FloorWalkerLoader (the last stupid class in the map scene)
Finally have got rid of mysterious event spaghetti :zap:
The flow of initialization is now so clear :smile:
The `init` method of MapScene seems a sort of untestable big function for the moment,
but it isn't such a problem because it can easily be split into small
methods by quite normal ways of refactoring.
It was really wrong to take the approach like "Everything is component".
That was a horribly wrong principle. Maybe I was being influenced a bit by
the approach of Polymer. I think they've been saying things like
"Everything is element". I understand it is possible to implement
everything as element. For example, an ajax call can be expressed as
HTML element like <x-ajax method="get" url="/data/floor/8" ... />. But
now I think this is a wrong approach. This kind of expression of things
make everything so independent and isolated from each other, that's a
good point, but the drawback is that it loses the view of flow of serial
processes. For example, about the things like ajax loading, it's
important to know who invokes the call and who handles the call. These
kind of things are actually well expressed by normal imperative
JavaScript code sequence, not by the isolated HTML element expression.
Actually if you isolate such things into custom dom element, then the
flow around the ajax call gets really hard to follow (that is exactly
what I've exprienced in this repository for a while). So therefore
it's far better to just implement these kind of things as you've been
doing since long time ago.